reorganizing a bit. - created a views package to house activities - created a service package to house services (firebase and ES)
