+++ app/src/main/java/com/bookyrself/bookyrself/views/CalendarActivity.java
@@ -1,11 +1,6 @@
-package com.bookyrself.bookyrself;
+package com.bookyrself.bookyrself.views;
 
-import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
+import com.bookyrself.bookyrself.R;
 
 public class CalendarActivity extends MainActivity {
