+++ app/src/main/java/com/bookyrself/bookyrself/views/MessagesActivity.java
@@ -1,16 +1,7 @@
-package com.bookyrself.bookyrself;
+package com.bookyrself.bookyrself.views;
 
-import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
-
+import com.bookyrself.bookyrself.R;
 import com.firebase.ui.auth.AuthUI;
-import com.google.android.gms.auth.api.Auth;
-import com.google.firebase.FirebaseApp;
-import com.google.firebase.auth.FirebaseAuth;
 
 public class MessagesActivity extends MainActivity {
