+++ app/src/main/res/layout/activity_calendar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.bookyrself.bookyrself.CalendarActivity">
+    tools:context="com.bookyrself.bookyrself.views.CalendarActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
