+++ app/src/main/res/layout/activity_event_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.bookyrself.bookyrself.MessagesActivity">
+    tools:context="com.bookyrself.bookyrself.views.MessagesActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
