+++ app/src/main/res/layout/activity_search.xml
@@ -5,7 +5,7 @@
     android:id="@+id/search_activity_root"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.bookyrself.bookyrself.SearchActivity">
+    tools:context="com.bookyrself.bookyrself.views.SearchActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/appBarLayout"
