+++ app/src/main/res/layout/activity_search_options.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="com.bookyrself.bookyrself.SearchActivity">
+    tools:context="com.bookyrself.bookyrself.views.SearchActivity">
 
     <Button
         android:id="@+id/search_btn2"
